requirements.txt
streamlit
requests
fastapi
uvicorn
google-generativeai
python-dotenv
